# api/swagger.yaml
---
swagger: "2.0"
info:
  version: "1.0.0"
  title: "API for kogger"
basePath: "/api"
tags:
- name: "Logs"
  description: "Endpoints for logs"
paths:
  /namespaces:
    get:
      tags:
      - "Logs"
      summary: "Get all namespaces"
      operationId: "getNamespaces"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "All namespaces retrieved successfully"
          schema:
            $ref: "#/definitions/NamespaceResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Not Found"
  /resources:
    get:
      tags:
      - "Logs"
      summary: "Get all resources in a namespace"
      operationId: "listResources"
      produces:
      - "application/json"
      parameters:
      - name: "namespace"
        in: "query"
        description: "Name of the namespace"
        required: true
        type: "string"
        x-exportParamName: "Namespace"
      responses:
        "200":
          description: "All resources in the namespace retrieved successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ResourcesList"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Not Found"
  /resources/{namespace}:
    get:
      tags:
      - "Logs"
      summary: "Get all resources in a namespace"
      operationId: "getResource"
      produces:
      - "application/json"
      parameters:
      - name: "namespace"
        in: "path"
        description: "Name of the namespace"
        required: true
        type: "string"
        x-exportParamName: "Namespace"
      - name: "resourceType"
        in: "query"
        description: "Type of the resource"
        required: true
        type: "string"
        x-exportParamName: "ResourceType"
      - name: "name"
        in: "query"
        description: "Name of the resource"
        required: true
        type: "string"
        x-exportParamName: "Name"
      responses:
        "200":
          description: "All resources in the namespace retrieved successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ResourcesList"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Not Found"
  /logs/{namespace}:
    get:
      tags:
      - "Logs"
      summary: "Get all logs"
      operationId: "getLogs"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "All logs retrieved successfully"
          schema:
            $ref: "#/definitions/LogResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Not Found"
definitions:
  NamespaceResponse:
    type: "object"
    required:
    - "namespaces"
    properties:
      namespaces:
        type: "array"
        items:
          $ref: "#/definitions/NamespaceInList"
    example:
      namespaces:
      - resourcesPath: "resourcesPath"
        name: "name"
      - resourcesPath: "resourcesPath"
        name: "name"
  ResourcesList:
    type: "object"
    required:
    - "namespace"
    - "resources"
    properties:
      namespace:
        type: "string"
        description: "Namespace of the resource"
      resources:
        type: "array"
        items:
          $ref: "#/definitions/ResourcesByType"
    example:
      namespace: "namespace"
      resources:
      - resources:
        - name: "name"
        - name: "name"
        resourceType: "pod"
      - resources:
        - name: "name"
        - name: "name"
        resourceType: "pod"
  ResourcesByType:
    type: "object"
    required:
    - "resourceType"
    - "resources"
    properties:
      resourceType:
        type: "string"
        example: "pod"
        description: "Type of the resource"
      resources:
        type: "array"
        items:
          $ref: "#/definitions/ResourcesInList"
    example:
      resources:
      - name: "name"
      - name: "name"
      resourceType: "pod"
  ResourcesInList:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
        description: "Name of the resource"
    example:
      name: "name"
  NamespaceInList:
    type: "object"
    required:
    - "name"
    - "resourcesPath"
    properties:
      name:
        type: "string"
        description: "Name of the namespace"
      resourcesPath:
        type: "string"
        description: "Api path to get resources for the namespace"
    example:
      resourcesPath: "resourcesPath"
      name: "name"
  LogResponse:
    type: "object"
    properties:
      Logs:
        type: "array"
        items:
          $ref: "#/definitions/Logs"
    example:
      Logs:
      - node: "node"
        pod: "pod"
        namespace: "namespace"
        logs: "logs"
        status: "status"
      - node: "node"
        pod: "pod"
        namespace: "namespace"
        logs: "logs"
        status: "status"
  Logs:
    type: "object"
    required:
    - "logs"
    - "namespace"
    - "node"
    - "pod"
    - "status"
    properties:
      namespace:
        type: "string"
        description: "Namespace of the pod"
      pod:
        type: "string"
        description: "Name of the pod"
      status:
        type: "string"
        description: "Status of the pod"
      node:
        type: "string"
        description: "Node where the pod is running"
      logs:
        type: "string"
        description: "Logs of the pod"
    example:
      node: "node"
      pod: "pod"
      namespace: "namespace"
      logs: "logs"
      status: "status"
  Response:
    type: "object"
    required:
    - "message"
    - "token"
    properties:
      token:
        type: "string"
      message:
        type: "string"