# koggerserver/routers.go
/*
 * API for kogger
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package koggerserver

import (
	"fmt"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

type Route struct {
	Name                string
	Method              string
	Pattern             string
	HandlerFunc         http.HandlerFunc
	RequiredPermissions []string
}

type Routes []Route

func NewRouter() *mux.Router {
	router := mux.NewRouter().StrictSlash(true).UseEncodedPath()
	for _, route := range routes {
		var handler http.Handler
		handler = route.HandlerFunc
		handler = Logger(handler, route.Name)

		router.
			Methods(route.Method).
			Path(route.Pattern).
			Name(route.Name).
			Handler(handler)
	}

	return router
}

func Index(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	fmt.Fprintf(w, "{\"name\": \"API for kogger\",\"version\": \"1.0.0\"}")
	w.WriteHeader(http.StatusOK)
}

func Healthz(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusOK)
}

var routes = Routes{
	Route{
		"Index",
		"GET",
		"/api/",
		Index,
		[]string{},
	},

	Route{
		"Healthz",
		"GET",
		"/healthz",
		Healthz,
		[]string{},
	},

	Route{
		"GetLogs",
		strings.ToUpper("Get"),
		"/api/logs/{namespace}",
		GetLogs,
		[]string{  },
	},

	Route{
		"GetNamespaces",
		strings.ToUpper("Get"),
		"/api/namespaces",
		GetNamespaces,
		[]string{  },
	},

	Route{
		"GetResource",
		strings.ToUpper("Get"),
		"/api/resources/{namespace}",
		GetResource,
		[]string{  },
	},

	Route{
		"ListResources",
		strings.ToUpper("Get"),
		"/api/resources",
		ListResources,
		[]string{  },
	},
}