# models/model_namespace_response.go
/*
 * API for kogger
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type NamespaceResponse struct {

	Namespaces []NamespaceInList `json:"namespaces"`
}