# models/model_resources_by_type.go
/*
 * API for kogger
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type ResourcesByType struct {

	// Type of the resource
	ResourceType string `json:"resourceType"`

	Resources []ResourcesInList `json:"resources"`
}