# models/model_resources_list.go
/*
 * API for kogger
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type ResourcesList struct {

	// Namespace of the resource
	Namespace string `json:"namespace"`

	Resources []ResourcesByType `json:"resources"`
}